Consolidate DAG traits
