+++ casper/src/test/scala/coop/rchain/casper/util/DagOperationsTest.scala
@@ -4,11 +4,11 @@ import cats.{Id, Monad}
 import coop.rchain.casper.helper.BlockGenerator._
 import coop.rchain.casper.helper.{BlockDagStorageFixture, BlockGenerator}
 import coop.rchain.casper.protocol.BlockMessage
+import coop.rchain.dag.DagOps
 import coop.rchain.shared.scalatestcontrib._
 import coop.rchain.models.BlockMetadata
 import coop.rchain.models.BlockHash.BlockHash
 import coop.rchain.models.Validator.Validator
-import coop.rchain.shared.DagOps
 import monix.eval.Task
 import org.scalatest.{FlatSpec, Matchers}
